minor fixes
The About Us page linked to archive/Venue.php but that page is not in an archive folder but at the root.

The Ads page still connects to the State Fair ad form. This cannot be totally fixed until a new ad form can be created for How to Succeed.

The How to Succeed cast page had the Gondoliers synopsis.  It now has the correct synopsis.